The Set of downloaded ids was created together with the duplication validation in the function save_episodes

+++ main.py
@@ -1,49 +1,67 @@
-import pandas as pd 
+import pandas as pd
 import requests as requests
 import json
-from pathlib import Path    
+from pathlib import Path
+import os
 
-# Installin Competitions.csv
+# Installing Competitions.csv
 df = pd.read_csv('Competitions.csv')
 #print(df.head())
 
 # Filter rows where id == 17592
-id = df[df['Slug'] == 'connectx']['Id']
+competition_id = df[df['Slug'] == 'connectx']['Id']
 
 # Load the episodes dataset
 episodes_df = pd.read_csv('Episodes.csv')  # adjust the path if needed
 
 # Filter rows where competitionId == 17592
-filtered_episodes = episodes_df[episodes_df['CompetitionId'] == id.item()]
+filtered_episodes = episodes_df[episodes_df['CompetitionId'] == competition_id.item()]
 
 # Get the list of episode IDs
 EpisodeId = filtered_episodes['Id'].tolist()
 
-# Create outpu_folder for episods results
-OUTPUT_DIR = Path("Episods_output")
+# Create outpu_folder for episodes results
+OUTPUT_DIR = Path("Episodes_output")
 
 LOWEST_SCORE_THRESH = 2000
-EPISODE_LIMIT_SIZE = 10
+EPISODE_LIMIT_SIZE = 15
 
-# Function for  saving the episods results in json format
-def saveEpisode(EpisodeId:int):
+# Get the list of ids' of all files from folder Episodes_output
+downloaded_files = {f for f in os.listdir(OUTPUT_DIR)
+                    if f.endswith('.json')}
+downloaded_id = {os.path.splitext(f)[0] for f in downloaded_files}
+
+
+# Function for  saving the episodes results in json format
+def save_episode(episode_id: int):
+    # Duplicates verification
+    filepath = OUTPUT_DIR / f"{episode_id}.json"
+    if filepath.exists():
+        print(f"Episode {episode_id} already downloaded. Skipping.")
+        return
+
     # Create URL for request
-    GET_URL = f"https://www.kaggleusercontent.com/episodes/{EpisodeId}.json"
+    get_url = f"https://www.kaggleusercontent.com/episodes/{episode_id}.json"
     # request
-    re = requests.get(GET_URL)
+    re = requests.get(get_url)
     if re.status_code == 200:
         try:
             # save replay
             replay = re.json()
             OUTPUT_DIR.mkdir(exist_ok=True)
-            with open(OUTPUT_DIR / f'{EpisodeId}.json', 'w') as f:
+            with open(OUTPUT_DIR / f'{episode_id}.json', 'w') as f:
                 json.dump(replay, f)
-            print(f"Episode {EpisodeId} successfully saved.")
-        except Exception as e:
-            print(f"JSON decode error for episode {EpisodeId}: {e}")
-    else:
-        print(f"Request error for episode {EpisodeId}: status {re.status_code}")
+            print(f"Episode {episode_id} successfully saved.")
 
-# Launch and Save the first 10 episodes  
+            # Append new ids in the downloaded list of episodes
+            downloaded_id.add(str(episode_id))
+
+        except Exception as e:
+            print(f"JSON decode error for episode {episode_id}: {e}")
+    else:
+        print(f"Request error for episode {episode_id}: status {re.status_code}")
+
+
+# Launch and Save the first 15 episodes
 for eid in EpisodeId[:EPISODE_LIMIT_SIZE]:
-    saveEpisode(eid)
+    save_episode(eid)
